Move over to `service-client`
Remove all references to `utilities`. There are a couple of JSON-loading
functions that still only exist in `utilities`; these were intended to
be shared with acs-auth but in fact that code has its own
implementation. Copy them in for now; they should be replaced with the
YAML code from service-setup at some point, assuming we still need to
load dumps on every startup at all.