Batch poll requests in the queue
When we are polling serially we need to queue poll requests. We need to
limit the queue size for sanity reasons; if we're overrunning the queue
the device isn't keeping up with the poll requests. However, with addrs
queued individually we were quickly overrunning the queue on devices
with many addrs.

Instead, queue each poll request in full. This means processing a queue
item will take longer, and there's more chance of hitting the 30s
backstop timeout, but we have to enforce limits somehow.