Drive watch_powerset from watch_subclasses
If this method is `watch_powerset` then it should be watching a
powerset. Without expicit PS support in the ConfigDB this means we need
to track the members of the PS via watch_subclasses.

For the current use case (the Auth service) this is probably more
convenient; it reduces the number of well-known UUIDs required, for a
start. I think in practice the groups used as principal or permission
groups will be _Role_ / _Composite permission_ rather than the
powersets, but I don't think at this point I want to enforce that as a
restriction.

Include members-of-members as an additional option in case it's useful.