Rework groups API to only support migration
Migration here is a little tricky; we need to migrate the members of the
Auth groups into the ConfigDB classes, but we can't do that until we
have both new services running and service-setup has migrated the dumps.

Although it stinks a little, keep a limited version of the old Auth
groups API for now for migration. This will only support listing and
removing groups. I think this is cleaner than trying to arrange for
service-setup to have direct database access.

Rev the Auth spec version to 2.0.0. This API spec isn't complete yet but
we need to start publishing it (from dev builds only) immediately.