Cache the ACL for each principal
Cache the sequence giving the full ACL for each principal for half an
hour. The HTTP API picks one item off the front of the sequence so this
will avoid recalculating unnecessarily; if we haven't had any requests
about a given principal for half an hour they've probably gone away for
now.