Update grants seq when we make a change
I'm not going to attempt to drive this from Pg NOTIFY, it didn't work
well when I did that for the Directory. So we have to drive updates from
what we know we've done. In principle we should be able to track changes
without going back to the database from the requests made and their
outcomes, but for now just refetch the whole list.