Migrate connection information from EA configs
The Edge Scout work, and potentially the new admin interface and other
possible work, will require that connections are identified individually
in the ConfigDB. This means existing connection information needs to be
extracted from the Edge Agent configs and put into its own entries.

We also need ConfigDB objects for drivers. The built-in and stock
drivers will need to be registered by a dump; this will replace the
defaults in the Helm chart eventually.