Update existing dumps to specify plural targets
Most targets appear to be singular so make that the default. We still
need to require dump format version 2 so we can be sure this isn't
ambiguous. (In fact many of these grants should be plural in some form,
such as the ReadConfig grants, but we can't handle that yet.)