Support loading dynamic dumps
The only thing preventing us from doing most of this work from a dump is
the dynamic UUIDs. If we extend the DumpLoader to accept dynamic UUIDs
(keyed under `Local`) then we can load all this configuration from
dumps. This is simpler than trying to replicate the insert-or-update
logic which the ordinary API doesn't support.