Install Helm charts via the dumps
We need dynamic UUIDs for roles and charts; create these as an initial
service-setup step before loading the dumps. Then we can simply
reference them from the dumps using the `Local` prefix.

Generate the existing ServiceConfig entries for compatibility. Probably
the services want updating to use the master list directly, or we want a
_Well-known name_ app attached directly to each UUID, or something. The
HelmTemplate ServiceConfig only existed to preserve the UUIDs and
doesn't need keeping.