Use dumps to set up cluster manager
This was not so simple: in principle we will create arbitrary git repo
mirrors, so that step must still be done in code. Change the search
strategy here to explicitly resolve the requested repo by name; this is
probably more robust (against removing and recreating a mirror in the
values file) as well as simpler.

We must still create the Helm charts repo in the dumps; that is a UUID
we depend on.