Simplify the NotifyV2 interface
Instead of subclassing the NotifyV2 class, follow the Express pattern
and treat it as a router object with a separate class for the handlers.
This means that NotifyV2 doesn't need to get involved with authz or the
model or any of that.

Rework the ConfigDB notify interface to use the new API.