Rework track_identities using Response
I';m trying to remove some of the duplication between the HTTP and
notify APIs. Towards this end, rework track_identities so it returns an
Optional sequence of Responses; this can then be interpreted by the two
APIs in their respective ways.

Ideally I will get to the point of having an Optional seq of Responses
for each endpoint, and then completely generic protocol translators.