Handle notify errors better
When a notify error update came in we were sitting in `switchMap`
waiting for a new WS to talk to. This is not a helpful reaction. Make
sure we pass the error all the way down the sequence, and that cached
seqs reset properly so we can retry. This informs the client and allows
them to use `retry` as needed.