Implement direction relationship stores
I'm not overly happy with this commit, but I think it's an improvement at least. The direct relationship store is now split into two (members and subclasses). We could have one large store if I knew how to combine all the rx pipes. This will need to be used with a cluster that is upgraded beyond 4.0.0-dev.4 since the notify interface was after this release.