Rename the Keycloak bootstrap user
I am having a lot of trouble creating the admin@ user account; Keycloak
keeps returning 409 even though it later denies the account exists. I am
wondering if this is because of a conflict with the initial 'temporary'
bootstrap account username. Rename to avoid this; at the same time
change the password, this isn't linked to admin@ any more.