Create admin@ linked to Kerberos
The Keycloak user-create endpoint is consistently returning 409 when we
attempt to create an account for the admin user. I think this may be
because we have federation set up to Kerberos and Keycloak refuses to
support non-Kerberos users in this realm. IMHO this is desirable in any
case.

We don't have a real human driving to perform the first-login flow, so
go round the houses a bit. First we request a token 'on behalf of'
admin@; this will fail because we haven't set up the profile, but will
create the user on the Keycloak side. Then we can go in with our user
management credentials and finish the setup.