We can't remove the grafana-internal admin user
If we don't specify an admin account secret Grafana generates one with
the username `admin`, which conflicts with the Kerberos account we are
trying to use via OpenID. Replace the Secret but change the username and
give it a password. It won't be possible to log in as this user in any
case as the login dialog is disabled.