Try to get OpenID roles visible in Grafana
It seems like the only way to have the default install set up admin@ as
a Grafana admin is for Grafana to defer to Keycloak completely for role
information. This means that admin@ must also be able to manage the
roles assigned to other OpenID users; this information can no longer be
managed in Grafana.

* Allow assigning multiple default and admin roles.
* Allow assigning roles within builtin clients.
* Assign admin@ the roles needed to be a realm admin in Keycloak.
* Consistently use the key in openidClients as client ID.