Support custom drivers
The logic to use image references to deploy a driver will only work for
drivers distributed with ACS. Custom drivers won't have an image defined
in the edge-agent Helm chart and the code does not try to add the image
spec to the `image` section of the deployment.

The Helm chart has been changed to accept a full image spec in place of
an image reference. Use this feature to deploy the image specified in
the driver definition. The reference feature was a convenience for
creating these deployments by hand; for an automated system it's easier
to be specific.

Include the contents of the `deployment` property from the driver
definition in the driver-specific values. This is to allow configuration
of host path mounts, privileged containers and other advanced options.