Regenerate the Deployment entries
The structure of the values expected by the Helm chart has changed; this
means we must update the Deployment at this point. Because the Manager
performs updates incrementally and in the frontend we must duplicate the
update logic here.

We are still incrementally updating in that we pass those values not
under `drivers` through without recording them anywhere else.