If we sync the stores we can trust the results
We were generating incorrect configs due to not fetching the _Driver
information_ entries.

The stores provide a lot of useful functionality; it's foolish not to
use them. But currently they do still have a cache coherency issue
because we don't have global transactions. The config-generation code
already handles this by stopping and restarting the stores; trust the
results.