Auth dumps are authoritative about grants
When a dump loads grants for a given principal, it specifies the full
set of grants for that principal. All other existing grants are removed.
This means it's not possible to grant new permissions to users and
groups which are created by ACS but I think that can be worked around.