Check dump permissions in the model
If we are going to remove existing grants then the permission checks for
loading a dump depend on the existing data; this means they must be
performed in the model. We still have a race condition between freezing
our permitted lists and the DB transaction but I can't see how to avoid
that.