Grant edge krbkeys rights to all Edge Agents
An Edge Agent is created by a human user, and then the edge krbkeys
needs to place it in the correct groups. For now grant all edge krbkeys
access to write memberships of all Edge Agents; this is still more
restricted than before. Ideally we want a class of agents per cluster
and an individual grant for each edge krbkeys.