Cluster manager needs to classify clusters
As before, these objects are not created by the cluster manager but by
an admin. Part of the problem here is we are using the same object to
represent the cluster as a whole and the Git repo controlling it; the
classification here is of the repo. If they were separate objects then
the cluster manager could own the repo object; but then we would need to
be able to record the association to the cluster.