Build Shapes for ACL entries
This is very crude at present. The basic structure is that an ACL grant
is a triple with predicate `acl:readTriple`; the object of this is an
acl:ShexCondition identifying triples the subject may access.
(Annoyingly ShEx TripleConstraints are effectively constraints on a
node rather than on a triple as such).

These `readTriple` triples must be present in the default graph of the
dataset; for now I am ignoring other graphs, but if we use them it will
most likely be for hypothetical or otherwise 'quoted' triples so we
don't want them to be active permission grants.

A ShexCondition has three properties, `acl:subject`, `acl:predicate` and
`acl:object`, all ShEx expresssions. For now these are limited to:

    - value NodeConstraints
    - forwards TripleConstraints with cardinality 1

Since the Jena ShEx library won't (yet) compile ShEx expressions from
RDF, only from JSON or ShExC, we need to traverse the graph and build
the expressions by hand.

Our DatasetGraph builds the ACL for a principal, for now from scratch on
every request. This means collating all the Shapes, combining them into
a Schema, and recording a set of triples representing the ShexCondition
objects. It should be possible to use these to evaluate the ACLs.