Evaluate ACLs in the DatasetGraph
Convert the FPShapeBuilder into a builder class for FPShapeEvaluator;
this now holds the final ShexSchema. This removes the
Optional<ShexSchema> which is obviously a good thing.

At the moment I build the FPShapeEvaluator in get(); it doesn't feel
quite right building it in the constructor. This means another Optional,
unfortunately, since we're changing state again. A proper implementation
would use change-notify and maintain a single schema with all the shapes
in the graph in any case.

Implement find() on top of stream(), instead of the other way round. The
Iterator classes seem to be extremely limited.