Make services return 503 for any ServiceError
Rationalise our error returns in ServiceClient so that we only return
ServiceError for a problem with the remote service. Ordinary parameter
errors now throw other errors. This means that a ServiceError thrown
during processing of a request should always invoke a 503 response.

Remove the hack in Auth.fetch_acl to throw a ServiceError with 503 in.