If we retry updates, we must do so in parallel
The original logic here did not retry failed updates and left this to
the 10-minute reconcile loop. Changing it to retry individual actions
causes a retrying update to block all other update, including an update
to delete the cluster.

Group actions by cluster. Attach retry logic to each action and then
switch through the actions for each cluster. This ensures we abandon a
retrying action if we should be doing something else instead.