Rework cmdesc to use notify throughout
Instead of looking up addresses and CCL templates when we need them, and
relying on the HTTP caching, just track the whole lot via notify. This
simplifies a lot of the logic.

Remove redundant MQTT logic concerned with being a Sparkplug Node. We
don't need to do that any more.