Pass pull information into hooks
We need this to set the schema source, but it isn't entirely reliable.
An auto-pull branch can still be pushed to, and in that case the
information will not have come from the remote. But I am assuming that
in such a case an admin has done this for a good reason.